Moved DictMatrix to utils folder
